Fix indentation feature.
Previously the “allow indentation” feature did not just apply to a template, but also to the resulting text. This was no good if the ERB produced started with spaces, then those spaces would be removed.

This PR instead removes the whitespace from the template before it gets compiled.

I don’t advocate monkey patching Rails, if there’s a better place to hook in here, i’m all ears.