Giant Monkeypatch Refactor
Previously the behavior of the library was one giant hack. It is now split up to be two smaller sized hacks.

The markdown processing and rendering is now provided by ActionView Handlers interface which is a supported endpoint to hook into.

There are two monkeypatches. One to action_mailer removes the need to using blocks with formats in them in your mailer:

```
# This is not needed in your mailer.
do |format|
  format.text
  format.html
end
```

Of course you can also manually add this block and then the monkeypatch code is not needed (and never even triggered).

The second monkeypatch is used to allow `.md.erb` as an extension. If you want to avoid this you can instead use `.md` or `.md+erb`.

Layouts, which were previously a bolted-on hacked feature, now work out of the box as a result of using the action view handler. The ability to `Maildown.allow_indentation` actually works now, by stripping the template before the ERB is applied instead of after.

As a bonus you can now also use `.md` templates for your regular app view as well.