Fix compiled templates (#39)
Fixes a bug where text emails render fine the first time, but then are empty the second time.

Templates in rails will be “compiled”, this means their original source is removed and a compiled flag is added. This is good for performance, but it also means calling `dup` on the result of `original_each_template` the first time, gives a very different result than calling dup on `original_each_template` the second time since it has already been “compiled”.

I fixed this by storing the “text” template inside of the “html” template by setting an instance variable. This means that the exact same template for text emails will be used on the first, second, and every other call.