Patch for Rails 7
Documented in https://github.com/zombocom/maildown/issues/59. This class was removed from Rails 7.

This causes an error on this line when trying to alias a method that does not exist https://github.com/zombocom/maildown/blob/cbf5b51e0867eab1ec945d38b5335f42139b0322/lib/maildown/ext/action_view.rb#L7.

The solution presented was to re-add the method. However simply doing this with a `defined?(OptimizedFileSystemResolver)` causes a load via zeitwork and also results in a crash.

The presence of this code appears to do nothing in Rails 7. It's basically a no-op that allows us to boot without erroring.